octofit-tracker/backend/requirements.txt
Django==4.2.29
djangorestframework==3.14.0
django-cors-headers==4.5.0
